+++ src/Settings.ts
@@ -11,6 +11,16 @@ export class Settings {
 		{ serializer: serializerWithDefault<"dashed" | "colored">("dashed") }
 	);
 
+	public readonly jsonEscapeAssistantEnabled = new VsCodeSetting(
+		"hediet-power-tools.jsonEscapeAssistant.enabled",
+		{ serializer: serializerWithDefault<boolean>(true) }
+	);
+
+	public readonly stackFrameLineHighlighterEnabled = new VsCodeSetting(
+		"hediet-power-tools.stackFrameLineHighlighter.enabled",
+		{ serializer: serializerWithDefault<boolean>(true) }
+	);
+
 	public readonly customDefinitionsEnabled = new VsCodeSetting(
 		"hediet-power-tools.customDefinitions.enabled",
 		{ serializer: serializerWithDefault<boolean>(false) }
@@ -19,10 +29,5 @@ export class Settings {
 	public readonly debugAdapterLoggerEnabled = new VsCodeSetting(
 		"hediet-power-tools.debugAdapterLogger.enabled",
 		{ serializer: serializerWithDefault<boolean>(false) }
-	);
-
-	public readonly stackFrameLineHighlighterEnabled = new VsCodeSetting(
-		"hediet-power-tools.stackFrameLineHighlighter.enabled",
-		{ serializer: serializerWithDefault<boolean>(true) }
 	);
 }